backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
pandas
numpy
statsmodels
scikit-learn
